feature/dark-mode
new session modal
